remove unused func, add documentation
