+++ core/device_hooks/common_kernels.inc.cpp
@@ -53,7 +53,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/matrix/sellp_kernels.hpp"
 #include "core/matrix/sparsity_csr_kernels.hpp"
 #include "core/multigrid/amgx_pgm_kernels.hpp"
-#include "core/multigrid/mapping_kernels.hpp"
 #include "core/preconditioner/isai_kernels.hpp"
 #include "core/preconditioner/jacobi_kernels.hpp"
 #include "core/reorder/rcm_kernels.hpp"
@@ -1211,18 +1210,6 @@ GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
 namespace amgx_pgm {
 
 
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_AMGX_PGM_RESTRICT_APPLY_KERNEL(ValueType, IndexType)
-GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_AMGX_PGM_RESTRICT_APPLY_KERNEL);
-
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_AMGX_PGM_PROLONGATE_APPLY_KERNEL(ValueType, IndexType)
-GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_AMGX_PGM_PROLONGATE_APPLY_KERNEL);
-
 template <typename IndexType>
 GKO_DECLARE_AMGX_PGM_MATCH_EDGE_KERNEL(IndexType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
@@ -1292,25 +1279,6 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_IMPLICIT_RESIDUAL_NORM_KERNEL);
 
 
 }  // namespace implicit_residual_norm
-
-
-namespace mapping {
-
-
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_MAPPING_APPLYADD_KERNEL(ValueType, IndexType)
-GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MAPPING_APPLYADD_KERNEL);
-
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_MAPPING_ADVANCED_APPLYADD_KERNEL(ValueType, IndexType)
-GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MAPPING_ADVANCED_APPLYADD_KERNEL);
-
-
-}  // namespace mapping
 }  // namespace GKO_HOOK_MODULE
 }  // namespace kernels
 }  // namespace gko
