+++ dpcpp/CMakeLists.txt
@@ -31,7 +31,6 @@ target_sources(ginkgo_dpcpp
     matrix/sellp_kernels.dp.cpp
     matrix/sparsity_csr_kernels.dp.cpp
     multigrid/amgx_pgm_kernels.dp.cpp
-    multigrid/mapping_kernels.dp.cpp
     preconditioner/isai_kernels.dp.cpp
     preconditioner/jacobi_kernels.dp.cpp
     reorder/rcm_kernels.dp.cpp
