+++ include/ginkgo/ginkgo.hpp
@@ -88,8 +88,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/sparsity_csr.hpp>
 
 #include <ginkgo/core/multigrid/amgx_pgm.hpp>
-#include <ginkgo/core/multigrid/mapping.hpp>
-#include <ginkgo/core/multigrid/multigrid_base.hpp>
+#include <ginkgo/core/multigrid/multigrid_level.hpp>
 
 #include <ginkgo/core/preconditioner/ic.hpp>
 #include <ginkgo/core/preconditioner/ilu.hpp>
