+++ reference/CMakeLists.txt
@@ -22,7 +22,6 @@ target_sources(ginkgo_reference
     matrix/sellp_kernels.cpp
     matrix/sparsity_csr_kernels.cpp
     multigrid/amgx_pgm_kernels.cpp
-    multigrid/mapping_kernels.cpp
     preconditioner/isai_kernels.cpp
     preconditioner/jacobi_kernels.cpp
     reorder/rcm_kernels.cpp
