+++ reference/components/fill_array.cpp
@@ -30,10 +30,10 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#include <numeric>
-
-
 #include "core/components/fill_array.hpp"
+
+
+#include <numeric>
 
 
 namespace gko {
